Move CrashHandler to separate package
