+++ .circleci/test_modules.txt
@@ -13,4 +13,5 @@ permissions:testDebug
 settings:testDebug
 maps:testDebug
 errors:testDebug
+crash-handler:testDebug
 entities:testDebug
