+++ collect_app/src/main/java/org/odk/collect/android/activities/CrashHandlerActivity.kt
@@ -1,7 +1,7 @@
 package org.odk.collect.android.activities
 
 import android.os.Bundle
-import org.odk.collect.androidshared.ui.CrashHandler
+import org.odk.collect.crash_handler.CrashHandler
 import org.odk.collect.strings.localization.LocalizedActivity
 
 class CrashHandlerActivity : LocalizedActivity() {
