+++ settings.gradle
@@ -32,4 +32,5 @@ apply from: 'secrets.gradle'
 if (getSecrets().getProperty('MAPBOX_DOWNLOADS_TOKEN', '') != '') {
     include ':mapbox'
 }
+include ':crash-handler'
 include ':entities'
